+++ lib/giffy_dialog.dart
@@ -1,7 +1,6 @@
 library giffy_dialog;
 
-export 'src/giffy_dialog.dart';
-export 'src/giffy_bottom_sheet.dart';
+export 'src/giffy_modal.dart' show GiffyDialog, GiffyBottomSheet;
 export 'src/entry_animation.dart' show EntryAnimation;
 export 'package:rive/rive.dart';
 export 'package:lottie/lottie.dart';
